Remove redundant Application.Tests, consolidate into WorkflowInstanceTests
Delete 5 test files (EndEventTests, StartEventTests, ErrorHandlingTests,
ScriptTaskTests, TaskActivityTests) that duplicated coverage already
provided by WorkflowInstanceTests + Domain.Tests. Move 7 unique
activity-level snapshot/timestamp tests into WorkflowInstanceTests.

Application.Tests: 76 → 45 tests. All 202 tests pass.
